GOV-14369 Ability to limit Ownership Lookup field type to single responsibility

+++ d360.web/Controllers/V2/ResponsibilitiesController.cs
@@ -530,6 +530,19 @@ namespace d360.web.Controllers.V2
                         });
                         continue;
                     }
+
+                    string ownershipLookupMessage = ResponsibilityRepository.GetResponsibilityTypeUsedInOwnershipLookupMessage(responsibility, assetType);
+                    if (ownershipLookupMessage != "")
+                    {
+                        results.Add(new ResponsibilityTypeAllocationResponseModel()
+                        {
+                            AssetTypeUid = allocation.AssetTypeUid,
+                            Message = ownershipLookupMessage,
+                            Success = false
+                        });
+                        continue;
+                    }
+
                     results.Add(await ResponsibilityRepository.DeleteAllocation(responsibility, assetType, model.Cascade));
                 }
